+++ src/nco/nco_var_utl.c
@@ -1,4 +1,4 @@
-/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_var_utl.c,v 1.132 2009-05-01 22:31:24 zender Exp $ */
+/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_var_utl.c,v 1.133 2009-05-02 22:22:30 zender Exp $ */
 
 /* Purpose: Variable utilities */
 
@@ -14,7 +14,7 @@ nco_cpy_var_dfn /* [fnc] Copy variable metadata from input to output file */
  const int out_id, /* I [id] netCDF output file ID */
  const int rec_dmn_id, /* I [id] Input file record dimension ID  */
  const char * const var_nm, /* I [sng] Input variable name */
- const int *cnk_sz, /* I [nbr] Chunk sizes */
+ const size_t *cnk_sz, /* I [nbr] Chunk sizes */
  const int dfl_lvl) /* I [enm] Deflate level [0..9] */
 {
   /* Purpose: Copy variable metadata from input netCDF file to output netCDF file
@@ -97,7 +97,7 @@ nco_cpy_var_dfn_lmt /* Copy variable metadata from input to output file */
  const char * const var_nm, /* I [sng] Input variable name */
  CST_X_PTR_CST_PTR_CST_Y(lmt_all_sct,lmt_all_lst), /* I [sct] Hyperslab limits */
  const int lmt_all_lst_nbr, /* I [nbr] Number of hyperslab limits */
- const int *cnk_sz, /* I [nbr] Chunk sizes */
+ const size_t *cnk_sz, /* I [nbr] Chunk sizes */
  const int dfl_lvl) /* I [enm] Deflate level [0..9] */
 {
   /* Purpose: Copy variable metadata from input netCDF file to output netCDF file
@@ -827,7 +827,7 @@ nco_var_dfn /* [fnc] Define variables and write their attributes to output file 
  const int nbr_dmn_ncl, /* I [nbr] Number of dimensions in list */
  const int nco_pck_map, /* I [enm] Packing map */
  const int nco_pck_plc, /* I [enm] Packing policy */
- const int *cnk_sz, /* I [nbr] Chunk sizes */
+ const size_t *cnk_sz, /* I [nbr] Chunk sizes */
  const int dfl_lvl) /* I [enm] Deflate level [0..9] */
 {
   /* Purpose: Define variables in output file, copy their attributes */
